use age classification to determine who gets what vitals

+++ db/migrations/20251226211005_patient_evaluation_score.ts
@@ -9,7 +9,7 @@ export function up(db: Kysely<DB>) {
   }, (qb) =>
     qb.addColumn(
       'score',
-      'integer',
+      'smallint',
       (col) => col.notNull().check(sql`score >= 0`),
     ))
 }
